fix imports and move python virtual env scripts inside standard provider
