+++ providers/src/airflow/providers/apache/beam/hooks/beam.py
@@ -38,7 +38,7 @@ from packaging.version import Version
 from airflow.exceptions import AirflowConfigException, AirflowException
 from airflow.hooks.base import BaseHook
 from airflow.providers.google.go_module_utils import init_module, install_dependencies
-from airflow.utils.python_virtualenv import prepare_virtualenv
+from airflow.providers.standard.utils.python_virtualenv import prepare_virtualenv
 
 if TYPE_CHECKING:
     import logging
