+++ providers/src/airflow/providers/standard/utils/python_virtualenv_script.jinja2
